gl20: add SeeAnyPressAnyJob
SeeAnyPressAnyJob is for quickly adding a new job, you can specify which XML you want to load and mark the screenpoints you want to add as sapa="true"
