+++ app/build.gradle
@@ -7,8 +7,8 @@ android {
         applicationId "com.mumu.android.joshautomation"
         minSdkVersion 26
         targetSdkVersion 29   /* do not modify this */
-        versionCode 105
-        versionName "2.03"
+        versionCode 106
+        versionName "2.04"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
