+++ app/src/main/res/raw/light_definitions.xml
@@ -53,48 +53,44 @@
    1.53 -> 1.54   (NOT OK, add 0.01)
    This number will be considered as floating number so don't use another dots
 -->
-<definitions version="1.0">
-
+<definitions version="1.01">
     <defset resolution="1080x2340">
-        <screenpoints name="pPreBattle1">
+        <screenpoints name="pTapScreen" sapa="true">
+            <screenpoint>231,231,231,0xFF,1114,1024,L</screenpoint>
+            <screenpoint>253,253,253,0xFF,1305,1014,L</screenpoint>
+        </screenpoints>
+
+        <screenpoints name="pSkipResult" sapa="true">
+            <screenpoint>77,76,77,0xFF,1256,862,L</screenpoint>
+            <screenpoint>236,236,236,0xFF,1186,880,L</screenpoint>
+        </screenpoints>
+
+        <screenpoints name="pAutoBattle" sapa="true">
+            <screenpoint>190,190,190,0xFF,2244,998,L</screenpoint>
+        </screenpoints>
+
+        <screenpoints name="pPreBattle1" sapa="true">
             <screenpoint>217,186,117,0xFF,1443,1025,L</screenpoint>
             <screenpoint>236,234,231,0xFF,1396,990,L</screenpoint>
         </screenpoints>
 
-        <screenpoints name="pPreBattle1_0">
+        <screenpoints name="pPreBattle1_0" sapa="true">
             <screenpoint>109,92,73,0xFF,1458,971,L</screenpoint>
             <screenpoint>243,242,240,0xFF,1394,975,L</screenpoint>
             <screenpoint>233,209,141,0xFF,1386,1017,L</screenpoint>
         </screenpoints>
 
-        <screenpoints name="pPreBattle2">
+        <screenpoints name="pPreBattle2" sapa="true">
             <screenpoint>109,92,73,0xFF,1288,978,L</screenpoint>
         </screenpoints>
 
-        <screenpoints name="pPreBattle3">
+        <screenpoints name="pPreBattle3" sapa="true">
             <screenpoint>109,92,73,0xFF,1293,990,L</screenpoint>
         </screenpoints>
-
-        <screenpoints name="pAutoBattle">
-            <screenpoint>190,190,190,0xFF,2244,998,L</screenpoint>
-        </screenpoints>
-
-        <screenpoints name="pTapScreen">
-            <screenpoint>231,231,231,0xFF,1114,1024,L</screenpoint>
-            <screenpoint>253,253,253,0xFF,1305,1014,L</screenpoint>
-        </screenpoints>
-
-        <screenpoints name="pSkipResult">
-            <screenpoint>77,76,77,0xFF,1256,862,L</screenpoint>
-            <screenpoint>236,236,236,0xFF,1186,880,L</screenpoint>
-        </screenpoints>
-
-
     </defset>
 
     <!-- Not support for 720p devices, this is a demo defset -->
     <defset resolution="720x1280">
         <screenpoint name="test">1,1,1,1,333,333,L</screenpoint>
     </defset>
-
 </definitions>
